Feature denvx

+++ src/main/java/com/example/projeto_aerj/beta/entities/pagamentos/Pagamento.java
@@ -5,21 +5,21 @@ import java.time.LocalDateTime;
 
 public class Pagamento {
 
-    private long id;
+    private int id;
     private LocalDateTime data;
     private BigDecimal valor;
 
-    public Pagamento(long id, LocalDateTime data, BigDecimal valor) {
+    public Pagamento(int id, LocalDateTime data, BigDecimal valor) {
         this.id = id;
         this.data = data;
         this.valor = valor;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
